requirements.txt
fastapi==0.103.1
gunicorn==21.2.0
uvicorn==0.23.2
boto3==1.28.53
python-decouple==3.8
python-multipart==0.0.6
bcrypt==4.0.1
asyncio==3.4.3
